+++ src/callbacks.h
@@ -29,7 +29,7 @@ void       *smart_tpl_iterator_value(void *userdata, void *iterator);
 void        smart_tpl_iterator_next(void *userdata, void *iterator);
 void       *smart_tpl_variable(void *userdata, const char *name, size_t size);
 const char *smart_tpl_to_string(void *userdata, void *variable);
-size_t      smart_tpl_to_numeric(void *userdata, void *variable);
+int         smart_tpl_to_numeric(void *userdata, void *variable);
 int         smart_tpl_to_boolean(void *userdata, void *variable);
 size_t      smart_tpl_size(void *userdata, void *variable);
 void*       smart_tpl_modifier(void *userdata, const char *name, size_t size);
